Update package structure in tests
