+++ src/test/java/net/smartcosmos/usermanagement/role/resource/DeleteRoleResourceTest.java
@@ -1,4 +1,4 @@
-package net.smartcosmos.extension.tenant.rest.resource.role;
+package net.smartcosmos.usermanagement.role.resource;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -8,7 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.web.servlet.MvcResult;
 
 import net.smartcosmos.cluster.userdetails.util.UuidUtil;
-import net.smartcosmos.extension.tenant.rest.resource.AbstractTestResource;
+import net.smartcosmos.test.AbstractTestResource;
 import net.smartcosmos.test.security.WithMockSmartCosmosUser;
 import net.smartcosmos.usermanagement.DevKitUserManagementService;
 import net.smartcosmos.usermanagement.role.dto.RoleResponse;
